Clean up user registration to match new team pattern
