+++ backend/busybeaver/src/main/java/org/opm/busybeaver/dto/Users/UsernameRegisterDto.java
@@ -1,0 +1,14 @@
+package org.opm.busybeaver.dto.Users;
+
+import jakarta.validation.constraints.*;
+
+public record UsernameRegisterDto(
+        @NotBlank(message = "Missing 'username' attribute to make a new user")
+        @Size(min = 3, max = 100, message = "Username must be 3 to 100 characters")
+        String username
+) {
+    @Override
+    public String username() {
+        return username;
+    }
+}
